# metric optimization: simply data structure used for metric datapoints and counter storage