--- a/metrics/src/counter.rs
+++ b/metrics/src/counter.rs
@@ -59,11 +59,7 @@ macro_rules! create_counter {
 #[macro_export]
 macro_rules! inc_counter {
     ($name:expr, $level:expr, $count:expr) => {
-        unsafe {
-            if log_enabled!($level) {
-                $name.inc($level, $count)
-            }
-        };
+        unsafe { $name.inc($level, $count) };
     };
 }