--- a/metrics/src/counter.rs
+++ b/metrics/src/counter.rs
@@ -32,7 +32,6 @@ pub struct CounterPoint {
 }
 
 impl CounterPoint {
-    #[cfg(test)]
     pub fn new(name: &'static str) -> Self {
         CounterPoint {
             name,