--- a/metrics/src/lib.rs
+++ b/metrics/src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::integer_arithmetic)]
 pub mod counter;
 pub mod datapoint;
-mod metrics;
+pub mod metrics;
 pub mod poh_timing_point;
 pub use crate::metrics::{flush, query, set_host_id, set_panic_hook, submit};
 use std::sync::{