--- a/metrics/src/metrics.rs
+++ b/metrics/src/metrics.rs
@@ -159,25 +159,12 @@ impl MetricsAgent {
         Self { sender }
     }
 
-    fn collect_points(
-        points_map: &mut HashMap<log::Level, (CounterMap, Vec<DataPoint>)>,
-    ) -> Vec<DataPoint> {
-        let points: Vec<DataPoint> = [
-            Level::Error,
-            Level::Warn,
-            Level::Info,
-            Level::Debug,
-            Level::Trace,
-        ]
-        .iter()
-        .filter_map(|level| points_map.remove(level))
-        .flat_map(|(counters, points)| {
-            let counter_points = counters.into_iter().map(|(_, v)| v.into());
-            points.into_iter().chain(counter_points)
-        })
-        .collect();
-        points_map.clear();
-        points
+    fn collect_points(points: &mut Vec<DataPoint>, counters: &mut CounterMap) -> Vec<DataPoint> {
+        let mut ret: Vec<DataPoint> = points.drain(..).collect();
+        for (_, v) in counters.drain() {
+            ret.push(v.into());
+        }
+        ret
     }
 
     fn write(
@@ -217,6 +204,7 @@ impl MetricsAgent {
 
         writer.write(points);
     }
+
     fn run(
         receiver: &Receiver<MetricsCommand>,
         writer: &Arc<dyn MetricsWriter + Send + Sync>,
@@ -225,7 +213,9 @@ impl MetricsAgent {
     ) {
         trace!("run: enter");
         let mut last_write_time = Instant::now();
-        let mut points_map = HashMap::<log::Level, (CounterMap, Vec<DataPoint>)>::new();
+        let mut points = Vec::<DataPoint>::new();
+        let mut counters = CounterMap::new();
+
         let max_points = write_frequency.as_secs() as usize * max_points_per_sec;
 
         loop {
@@ -235,7 +225,7 @@ impl MetricsAgent {
                         debug!("metrics_thread: flush");
                         Self::write(
                             writer,
-                            Self::collect_points(&mut points_map),
+                            Self::collect_points(&mut points, &mut counters),
                             max_points,
                             max_points_per_sec,
                             last_write_time,
@@ -245,17 +235,10 @@ impl MetricsAgent {
                     }
                     MetricsCommand::Submit(point, level) => {
                         log!(level, "{}", point);
-                        let (_, points) = points_map
-                            .entry(level)
-                            .or_insert((HashMap::new(), Vec::new()));
                         points.push(point);
                     }
-                    MetricsCommand::SubmitCounter(counter, level, bucket) => {
+                    MetricsCommand::SubmitCounter(counter, _level, bucket) => {
                         debug!("{:?}", counter);
-                        let (counters, _) = points_map
-                            .entry(level)
-                            .or_insert((HashMap::new(), Vec::new()));
-
                         let key = (counter.name, bucket);
                         if let Some(value) = counters.get_mut(&key) {
                             value.count += counter.count;
@@ -277,7 +260,7 @@ impl MetricsAgent {
             if now.duration_since(last_write_time) >= write_frequency {
                 Self::write(
                     writer,
-                    Self::collect_points(&mut points_map),
+                    Self::collect_points(&mut points, &mut counters),
                     max_points,
                     max_points_per_sec,
                     last_write_time,
@@ -289,19 +272,15 @@ impl MetricsAgent {
     }
 
     pub fn submit(&self, point: DataPoint, level: log::Level) {
-        if log_enabled!($level) {
-            self.sender
-                .send(MetricsCommand::Submit(point, level))
-                .unwrap();
-        }
+        self.sender
+            .send(MetricsCommand::Submit(point, level))
+            .unwrap();
     }
 
     pub fn submit_counter(&self, counter: CounterPoint, level: log::Level, bucket: u64) {
-        if log_enabled!($level) {
-            self.sender
-                .send(MetricsCommand::SubmitCounter(counter, level, bucket))
-                .unwrap();
-        }
+        self.sender
+            .send(MetricsCommand::SubmitCounter(counter, level, bucket))
+            .unwrap();
     }
 
     pub fn flush(&self) {