#### Problem

https://github.com/solana-labs/solana/issues/24319

Metric agent use a HashMap of tuple of Vec and HashMap to store datapoint and counters. However, these data is flattened into a vector of datapoints before submission. 

This extra levels of nesting is not necessary, and can be removed.

The following shows the benchmarks comparison between the old and new implementation.
new 
```
test bench_counter_submission   ... bench:     890,414 ns/iter (+/- 634,869)
test bench_datapoint_submission ... bench:     487,493 ns/iter (+/- 106,276)

```
old
```
test bench_counter_submission   ... bench:     956,648 ns/iter (+/- 352,442)
test bench_datapoint_submission ... bench:     571,490 ns/iter (+/- 598,148)
```

For datapoints, the new implementation  is **17%** faster. For counters new implementation is **7%** faster with new implementation.
And datapoint submission is **1.8x** faster than counters.


#### Summary of Changes

use a vector to store datapoints and hashmap to store counters.




Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->