@jeffwashington 2022-04-18
Is it possible to go 1 level higher and not construct the log string for log messages that are below the current level? Some of these could have {:?} logging perhaps?