@HaoranYi 2022-04-18
never mind. looks like the `datapoint!` already check this.