@HaoranYi 2022-04-27
Yes. Looks like it is around 3000. I will open another request to add this.
Thanks.

On Wed, Apr 27, 2022, 10:55 AM Jeff Washington (jwash) <
***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In metrics/src/metrics.rs
> <https://github.com/solana-labs/solana/pull/24447#discussion_r859970056>:
>
> > -            Level::Trace,
> -        ]
> -        .iter()
> -        .filter_map(|level| points_map.remove(level))
> -        .flat_map(|(counters, points)| {
> -            let counter_points = counters.into_iter().map(|(_, v)| v.into());
> -            points.into_iter().chain(counter_points)
> -        })
> -        .collect();
> -        points_map.clear();
> -        points
> +    fn collect_points(points: &mut Vec<DataPoint>, counters: &mut CounterMap) -> Vec<DataPoint> {
> +        let mut ret: Vec<DataPoint> = Vec::default();
> +        std::mem::swap(&mut ret, points);
> +        for (_, v) in counters.drain() {
> +            ret.push(v.into());
>
> Do you know how large 'counters' can be? It may be worth reserving the
> space in ret for adding all of counters before pushing.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/solana-labs/solana/pull/24447#pullrequestreview-955147650>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABVSJG3LEJEQLUGORJ3D2TVHFPN5ANCNFSM5TWNYMHQ>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>